+++ InfiniTAM/ITMLib/Utils/ITMVector.h
@@ -235,7 +235,7 @@ namespace ITMLib {
 		}
 
 		_CPU_AND_GPU_CODE_ inline Vector3<float> normalised() const {
-			float norm = 1.0f / sqrtf((float)(this->x * this->x + this->y * this->y + this->z * this->z));
+			float norm = 1.0f / sqrt((float)(this->x * this->x + this->y * this->y + this->z * this->z));
 			return Vector3<float>((float)this->x * norm, (float)this->y * norm, (float)this->z * norm);
 		}
